requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
httpx==0.27.2
pydantic==2.5.0
python-dotenv==1.0.0

# Needed for EmailStr in Pydantic models
email-validator==2.1.0.post1

# P0.1: RSA key management for RS256 JWT signing
cryptography==41.0.7

# unison-common package (local install)
../unison-common/dist/unison_common-0.1.0-py3-none-any.whl
